pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dl-alpha-bench"
version = "0.1.0"
description = "Leakage-safe deep learning alpha research benchmark for finance"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "laoyu17" }
]
keywords = ["quant", "deep-learning", "finance", "alpha", "walk-forward", "purged-cv"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
gui = ["PyQt6>=6.6"]
plot = ["matplotlib>=3.8"]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6"
]

[project.scripts]
dl-alpha-bench = "dl_alpha_bench.cli:main"
dl-alpha-bench-gui = "dl_alpha_bench.gui.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
